fix: change get file name method for authentication module
