--- backend/FileModule/File.py
@@ -1,9 +1,0 @@
-class File():
-    def __init__(self, filename):
-        self.__filename = filename
-
-    def setFilename(self, filename):
-        self.__filename = filename
-
-    def getFilename(self):
-        return self.__filename
